@@ -1,17 +1,18 @@
 {
   "name": "@dontcode/sandbox",
-  "version": "0.4.4",
+  "version": "0.5.0",
   "peerDependencies": {
-    "@angular/common": "^12.1.1",
-    "@angular/core": "^12.2.13",
-    "primeflex": "^3.1.0",
+    "@angular/common": "^13.3.2",
+    "@angular/core": "^13.3.2",
+    "@angular/router": "^13.3.2",
+    "primeflex": "^3.1.3",
     "primeicons": "^5.0.0",
-    "primeng": "^12.2.2",
-    "@dontcode/core": "^0.4.1",
-    "@dontcode/plugin-common": "^0.4.0",
-    "rxjs": "^6.6.7"
+    "primeng": "^13.3.2",
+    "@dontcode/core": "^0.5.1",
+    "@dontcode/plugin-common": "^0.5.0",
+    "rxjs": "^7.5.5"
   },
   "dependencies": {
-    "tslib": "^2.3.0"
+    "tslib": "^2.3.1"
   }
 }
