@@ -1,11 +1,14 @@
 {
   "name": "@dontcode/plugin-screen",
-  "version": "0.4.4",
+  "version": "0.5.0",
   "peerDependencies": {
-    "@angular/common": "^12.1.1",
-    "@angular/core": "^12.1.1",
-    "@dontcode/core": "^0.4.1",
-    "@dontcode/plugin-common": "^0.4.0",
-    "rxjs": "^6.6.7"
+    "@angular/common": "^13.3.2",
+    "@angular/core": "^13.3.2",
+    "@dontcode/core": "^0.5.1",
+    "@dontcode/plugin-common": "^0.5.0",
+    "rxjs": "^7.5.5"
+  },
+  "dependencies": {
+    "tslib": "^2.3.1"
   }
 }
