@@ -4,99 +4,83 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "nx": "nx",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "nx workspace-lint && ng lint",
-    "e2e": "ng e2e",
-    "affected:apps": "nx affected:apps",
-    "affected:libs": "nx affected:libs",
-    "affected:build": "nx affected:build",
-    "affected:e2e": "nx affected:e2e",
-    "affected:test": "nx affected:test",
-    "affected:lint": "nx affected:lint",
-    "affected:dep-graph": "nx affected:dep-graph",
-    "affected": "nx affected",
-    "format": "nx format:write",
-    "format:write": "nx format:write",
-    "format:check": "nx format:check",
-    "update": "nx migrate latest",
-    "dep-graph": "nx dep-graph",
-    "help": "nx help",
-    "postinstall": "ngcc --properties es2015 browser module main && node ./decorate-angular-cli.js",
-    "workspace-generator": "nx workspace-generator",
-    "run:all": "node node_modules/@angular-architects/module-federation/src/server/mf-dev-server.js"
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "start": "nx serve",
+    "build": "nx build",
+    "test": "nx test"
   },
   "private": true,
   "dependencies": {
-    "@angular-architects/module-federation-runtime": "^12.5.3",
-    "@angular/animations": "^12.2.13",
-    "@angular/cdk": "^12.2.13",
-    "@angular/common": "^12.2.13",
-    "@angular/compiler": "^12.2.13",
-    "@angular/core": "^12.2.13",
-    "@angular/forms": "^12.2.13",
-    "@angular/platform-browser": "^12.2.13",
-    "@angular/platform-browser-dynamic": "^12.2.13",
-    "@angular/router": "^12.2.13",
-    "@dontcode/core": "^0.4.8",
-    "@fortawesome/angular-fontawesome": "^0.10.1",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-regular-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@ngrx/component-store": "12.5.1",
-    "@nrwl/angular": "13.1.4",
-    "all-country-data": "^0.1.1",
+    "@angular-architects/module-federation-runtime": "^14.2.3",
+    "@angular/animations": "^13.3.2",
+    "@angular/cdk": "^13.3.2",
+    "@angular/common": "^13.3.2",
+    "@angular/compiler": "^13.3.2",
+    "@angular/core": "^13.3.2",
+    "@angular/forms": "^13.3.2",
+    "@angular/platform-browser": "^13.3.2",
+    "@angular/platform-browser-dynamic": "^13.3.2",
+    "@angular/router": "^13.3.2",
+    "@dontcode/core": "^0.5.2",
+    "@fortawesome/angular-fontawesome": "^0.10.2",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-regular-svg-icons": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@ngrx/component-store": "13.1.0",
+    "@nrwl/angular": "13.10.1",
+    "all-country-data": "^0.2.2",
     "async-mutex": "^0.3.2",
-    "broadcast-channel": "^4.5.0",
-    "core-js": "^3.19.1",
-    "dexie": "^3.2.0",
-    "flag-icon-css": "^4.1.5",
-    "i18n-iso-countries": "^7.0.0",
+    "broadcast-channel": "^4.10.0",
+    "core-js": "^3.21.1",
+    "dexie": "^3.2.1",
+    "flag-icon-css": "^4.1.7",
+    "i18n-iso-countries": "^7.4.0",
     "jsonpath-plus": "^6.0.1",
-    "primeflex": "^3.1.0",
+    "primeflex": "^3.1.3",
     "primeicons": "^5.0.0",
-    "primeng": "^12.2.2",
-    "rxjs": "^6.6.7",
+    "primeng": "^13.3.2",
+    "rxjs": "^7.5.5",
     "tslib": "^2.3.1",
-    "zone.js": "^0.11.4"
+    "zone.js": "0.11.5"
   },
   "devDependencies": {
-    "@angular-architects/module-federation": "^12.5.3",
-    "@angular-devkit/build-angular": "^12.2.13",
-    "@angular-eslint/eslint-plugin": "^12.6.1",
-    "@angular-eslint/eslint-plugin-template": "^12.6.1",
-    "@angular-eslint/template-parser": "^12.6.1",
-    "@angular/cli": "^12.2.13",
-    "@angular/compiler-cli": "^12.2.13",
-    "@angular/language-service": "^12.2.13",
-    "@nrwl/cypress": "^13.1.4",
-    "@nrwl/eslint-plugin-nx": "^13.1.4",
-    "@nrwl/jest": "^13.1.4",
-    "@nrwl/linter": "^13.1.4",
-    "@nrwl/workspace": "^13.1.4",
-    "@types/jest": "^27.0.2",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
+    "@angular-architects/module-federation": "^14.2.3",
+    "@angular-devkit/build-angular": "^13.3.2",
+    "@angular-eslint/eslint-plugin": "^13.2.0",
+    "@angular-eslint/eslint-plugin-template": "^13.2.0",
+    "@angular-eslint/template-parser": "^13.2.0",
+    "@angular/cli": "^13.3.2",
+    "@angular/compiler-cli": "^13.3.2",
+    "@angular/language-service": "^13.3.2",
+    "@nrwl/cli": "13.10.1",
+    "@nrwl/cypress": "13.10.1",
+    "@nrwl/eslint-plugin-nx": "13.10.1",
+    "@nrwl/jest": "13.10.1",
+    "@nrwl/linter": "13.10.1",
+    "@nrwl/workspace": "13.10.1",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^16.11.7",
+    "@typescript-eslint/eslint-plugin": "^5.18.0",
+    "@typescript-eslint/parser": "^5.18.0",
     "codelyzer": "^6.0.2",
-    "cypress": "^7.7.0",
-    "dotenv": "10.0.0",
-    "eslint": "7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-cypress": "^2.11.3",
-    "fake-indexeddb": "^3.1.3",
-    "jest": "^27.2.3",
-    "jest-preset-angular": "^10.0.1",
-    "ng-packagr": "12.2.5",
-    "postcss": "^8.3.9",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.1",
-    "prettier": "^2.3.2",
-    "ts-jest": "^27.0.5",
-    "ts-node": "^10.1.0",
-    "typescript": "^4.3.5"
+    "cypress": "^9.5.3",
+    "dotenv": "^16.0.0",
+    "eslint": "^8.13.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-cypress": "^2.12.1",
+    "fake-indexeddb": "^3.1.7",
+    "jest": "^27.5.1",
+    "jest-preset-angular": "^11.1.1",
+    "ng-packagr": "^13.3.0",
+    "nx": "13.10.1",
+    "postcss": "^8.4.12",
+    "postcss-import": "^14.1.0",
+    "postcss-preset-env": "^7.4.3",
+    "postcss-url": "^10.1.3",
+    "prettier": "^2.6.2",
+    "ts-jest": "^27.1.4",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.3"
   },
   "optionalDependencies": {
     "fsevents": "^2.3.2"
