@@ -1,85 +1,118 @@
-import {ChangeDetectionStrategy, ChangeDetectorRef, Component, NgZone, OnDestroy, OnInit} from "@angular/core";
-import {Subscription} from "rxjs";
-import {map} from "rxjs/operators";
-import {Change, ChangeType, DontCodeModel, DontCodeModelPointer} from "@dontcode/core";
-import {Router} from "@angular/router";
-import {MenuItem} from 'primeng/api';
-import {ChangeProviderService} from "../../shared/command/services/change-provider.service";
+import {
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  NgZone,
+  OnDestroy,
+  OnInit,
+} from '@angular/core';
+import { Subscription } from 'rxjs';
+import { map } from 'rxjs/operators';
+import {
+  Change,
+  ChangeType,
+  DontCodeModel,
+  DontCodeModelPointer,
+} from '@dontcode/core';
+import { Router } from '@angular/router';
+import { MenuItem } from 'primeng/api';
+import { ChangeProviderService } from '../../shared/command/services/change-provider.service';
 
 @Component({
   selector: 'dontcode-sandbox-menu',
   templateUrl: './menu.component.html',
   styleUrls: ['./menu.component.css'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class MenuComponent implements OnInit, OnDestroy {
-
-  menus:Array<MenuItem>;
-  templateMenus =new Array<MenuItem> (
-    {label:'Main Menu', items:[
-        {label:'Home', icon:'pi pi-home', routerLink:['/']},
-        {label:'Dev', icon:'pi pi-book', routerLink: ['dev']}
-      ]},
-    {label:'Application Menu', items:new Array<any>()}
+  menus: Array<MenuItem>;
+  templateMenus = new Array<MenuItem>(
+    {
+      label: 'Main Menu',
+      items: [
+        { label: 'Home', icon: 'pi pi-home', routerLink: ['/'] },
+        { label: 'Dev', icon: 'pi pi-book', routerLink: ['dev'] },
+      ],
+    },
+    { label: 'Application Menu', items: new Array<any>() }
   );
   subscriptions = new Subscription();
   runtime = false;
 
-  constructor(protected provider: ChangeProviderService,
-              private ref: ChangeDetectorRef, public router: Router,
-              public ngZone:NgZone) {
-    const config =(window as any).dontCodeConfig;
-    if( config?.runtime===true || config?.projectId!=null)
+  constructor(
+    protected provider: ChangeProviderService,
+    private ref: ChangeDetectorRef,
+    public router: Router,
+    public ngZone: NgZone
+  ) {
+    const config = (window as any).dontCodeConfig;
+    if (config?.runtime === true || config?.projectId != null)
       this.runtime = true;
-    this.menus = this.generateMenu ();
+    this.menus = this.generateMenu();
   }
 
   ngOnInit(): void {
-    this.handleEntitiesMenu (DontCodeModel.APP_ENTITIES, DontCodeModel.APP_ENTITIES_NAME_NODE, 'pi-ticket');
-    this.handleEntitiesMenu (DontCodeModel.APP_SCREENS, DontCodeModel.APP_SCREENS_NAME_NODE, 'pi-desktop');
+    this.handleEntitiesMenu(
+      DontCodeModel.APP_ENTITIES,
+      DontCodeModel.APP_ENTITIES_NAME_NODE,
+      'pi-ticket'
+    );
+    this.handleEntitiesMenu(
+      DontCodeModel.APP_SCREENS,
+      DontCodeModel.APP_SCREENS_NAME_NODE,
+      'pi-desktop'
+    );
   }
 
-  handleEntitiesMenu (entity:string, nameKey:string, icon:string) {
-   this.subscriptions.add (this.provider.receiveCommands (entity, nameKey).pipe (
-      map(command => {
-        this.updateMenuName (command, icon);
-        this.ref.detectChanges();
-      })
-    ).subscribe());
-    this.subscriptions.add (this.provider.receiveCommands (entity+"/?").pipe (
-      map(command => {
-        if (command.position.length>(entity.length+1))  // Avoid adding empty entities (received due to reset)
-        {
-          this.updateMenu (command, icon);
-          this.ref.detectChanges();
-        } else if (!command.value) {
-          // Reset all menus
-          this.getDynamicMenu().length=0;
-          this.ref.detectChanges();
-        }
-      })
-    ).subscribe());
+  handleEntitiesMenu(entity: string, nameKey: string, icon: string) {
+    this.subscriptions.add(
+      this.provider
+        .receiveCommands(entity, nameKey)
+        .pipe(
+          map((command) => {
+            this.updateMenuName(command, icon);
+            this.ref.detectChanges();
+          })
+        )
+        .subscribe()
+    );
+    this.subscriptions.add(
+      this.provider
+        .receiveCommands(entity + '/?')
+        .pipe(
+          map((command) => {
+            if (command.position.length > entity.length + 1) {
+              // Avoid adding empty entities (received due to reset)
+              this.updateMenu(command, icon);
+              this.ref.detectChanges();
+            } else if (!command.value) {
+              // Reset all menus
+              this.getDynamicMenu().length = 0;
+              this.ref.detectChanges();
+            }
+          })
+        )
+        .subscribe()
+    );
   }
 
-  getDynamicMenu (): Array<MenuItem> {
-      if( this.templateMenus[1].items)
-        return this.templateMenus[1].items;
-      else
-        return [];
+  getDynamicMenu(): Array<MenuItem> {
+    if (this.templateMenus[1].items) return this.templateMenus[1].items;
+    else return [];
   }
 
-  generateMenu (): Array<any> {
+  generateMenu(): Array<any> {
     // Create a new menu object to update UI
-    const ret= new Array<any>();
-    if( !this.runtime) {
-      this.templateMenus.forEach(value => {
+    const ret = new Array<any>();
+    if (!this.runtime) {
+      this.templateMenus.forEach((value) => {
         ret.push(value);
       });
     } else {
-      this.getDynamicMenu().forEach(value => {
+      this.getDynamicMenu().forEach((value) => {
         ret.push(value);
       });
-      if( ret.length===0) {
+      if (ret.length === 0) {
         ret.push(this.templateMenus[1]);
       }
     }
@@ -91,16 +124,16 @@ export class MenuComponent implements OnInit, OnDestroy {
     this.subscriptions.unsubscribe();
   }
 
-  gotoPage(page:string): void {
+  gotoPage(page: string): void {
     // ngZone is necessary as we are being called by a non angular component (kor-ui)
-    this.ngZone.run (() => {
+    this.ngZone.run(() => {
       this.router.navigate([page]);
-    })
+    });
   }
 
-  isActive(page:string):boolean {
+  isActive(page: string): boolean {
     const ret = this.router.isActive(page, true);
-//    console.log(page +' is active:'+ret);
+    //    console.log(page +' is active:'+ret);
     return ret;
   }
 
@@ -108,9 +141,11 @@ export class MenuComponent implements OnInit, OnDestroy {
     /*    if (position.startsWith(DontCodeModel.ROOT))
           position = position.substr(DontCodeModel.ROOT.length+1);*/
 
-    if (position.endsWith(DontCodeModel.APP_SCREENS_NAME_NODE))
-    {
-      position = position.substring(0, position.length-DontCodeModel.APP_SCREENS_NAME_NODE.length-1);
+    if (position.endsWith(DontCodeModel.APP_SCREENS_NAME_NODE)) {
+      position = position.substring(
+        0,
+        position.length - DontCodeModel.APP_SCREENS_NAME_NODE.length - 1
+      );
     }
     return position;
   }
@@ -120,14 +155,13 @@ export class MenuComponent implements OnInit, OnDestroy {
     const pos = this.findMenuPosOf(key);
     let menu;
 
-    if (pos===-1)
-    {
-      if( change.value?.name) {
-        menu= {
+    if (pos === -1) {
+      if (change.value?.name) {
+        menu = {
           routerLink: [key],
           label: change.value.name,
-          icon: 'pi ' + icon
-        }
+          icon: 'pi ' + icon,
+        };
       } else return;
     } else {
       menu = this.getDynamicMenu()[pos];
@@ -140,36 +174,40 @@ export class MenuComponent implements OnInit, OnDestroy {
       case ChangeType.DELETE:
         // These are handled at the name change level
         break;
-      case ChangeType.MOVE: {
-          if( pos!==-1) {
+      case ChangeType.MOVE:
+        {
+          if (pos !== -1) {
             this.getDynamicMenu().splice(pos, 1);
           }
           let beforeKeyPos = -1;
-          if(change.pointer) beforeKeyPos = this.findMenuPosOf(change.pointer.containerPosition+'/'+change.beforeKey);
+          if (change.pointer)
+            beforeKeyPos = this.findMenuPosOf(
+              change.pointer.containerPosition + '/' + change.beforeKey
+            );
 
-          if( beforeKeyPos!==-1)
+          if (beforeKeyPos !== -1)
             this.getDynamicMenu().splice(beforeKeyPos, 0, menu);
-          else
-            this.getDynamicMenu().push(menu);
-          }
+          else this.getDynamicMenu().push(menu);
+        }
         this.menus = this.generateMenu();
         break;
     }
   }
 
   private updateMenuName(command: Change, icon: string) {
-    const key = this.cleanPosition (DontCodeModelPointer.parentPosition(command.position)!);
-    const pos = this.findMenuPosOf (key);
-    const name=command.value;
+    const key = this.cleanPosition(
+      DontCodeModelPointer.parentPosition(command.position)!
+    );
+    const pos = this.findMenuPosOf(key);
+    const name = command.value;
 
     let menu;
-    if (pos===-1)
-    {
-      menu= {
+    if (pos === -1) {
+      menu = {
         routerLink: [key],
         label: command.value,
-        icon: 'pi ' + icon
-      }
+        icon: 'pi ' + icon,
+      };
     } else {
       menu = this.getDynamicMenu()[pos];
     }
@@ -178,25 +216,25 @@ export class MenuComponent implements OnInit, OnDestroy {
       case ChangeType.UPDATE:
       case ChangeType.RESET:
       case ChangeType.ADD:
-        if (pos!==-1) {
+        if (pos !== -1) {
           menu.label = name;
         } else {
           this.getDynamicMenu().push(menu);
         }
         break;
       case ChangeType.DELETE:
-        if (pos!==-1)
-          this.getDynamicMenu().splice(pos, 1);
+        if (pos !== -1) this.getDynamicMenu().splice(pos, 1);
         break;
-      case ChangeType.MOVE: {
-        // The move is handled at the menu level
-      }
+      case ChangeType.MOVE:
+        {
+          // The move is handled at the menu level
+        }
         break;
     }
     this.menus = this.generateMenu();
   }
 
-  private findMenuPosOf (key: string): number {
+  private findMenuPosOf(key: string): number {
     let pos = -1;
     this.getDynamicMenu().forEach((value, index) => {
       if (value.routerLink[0] === key) {
@@ -206,4 +244,3 @@ export class MenuComponent implements OnInit, OnDestroy {
     return pos;
   }
 }
-
