@@ -1,47 +1,51 @@
-import { Component, OnDestroy, OnInit } from "@angular/core";
-import { Change, CommandProviderInterface, DontCodeModelPointer, PreviewHandler} from "@dontcode/core";
-import {Subject} from "rxjs";
-import {takeUntil} from "rxjs/operators";
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import {
+  Change,
+  CommandProviderInterface,
+  DontCodeModelPointer,
+  PreviewHandler,
+} from '@dontcode/core';
+import { Subject, Subscription } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 
 @Component({
   selector: 'plugins-screen',
   templateUrl: './screen.component.html',
-  styleUrls: ['./screen.component.css']
+  styleUrls: ['./screen.component.css'],
 })
-export class ScreenComponent implements OnInit, OnDestroy, PreviewHandler {
+export class ScreenComponent implements OnDestroy, PreviewHandler {
+  unsubscriber = new Subscription();
 
-  unsubscriber = new Subject();
+  constructor() {}
 
-  constructor() {
-  }
-
-  ngOnInit(): void {
-  }
-
-  initCommandFlow(provider: CommandProviderInterface, pointer: DontCodeModelPointer) {
-//    console.log("InitCommandFlow");
-    provider.receiveCommands(pointer.position).pipe(
-      takeUntil(this.unsubscriber)
-    ).subscribe(change => {
-      console.log("Getting updates from ", pointer.position, " with value ", change.value);
-    });
+  initCommandFlow(
+    provider: CommandProviderInterface,
+    pointer: DontCodeModelPointer
+  ) {
+    //    console.log("InitCommandFlow");
+    this.unsubscriber.add(
+      provider.receiveCommands(pointer.position).subscribe((change) => {
+        console.log(
+          'Getting updates from ',
+          pointer.position,
+          ' with value ',
+          change.value
+        );
+      })
+    );
   }
 
   ngOnDestroy(): void {
-    this.forceUnsubscribe ();
+    this.forceUnsubscribe();
   }
 
-  forceUnsubscribe (): void {
+  forceUnsubscribe(): void {
     // unsubscribe to all observables
-//    console.log("Unsubscribe");
-    this.unsubscriber.next();
-    this.unsubscriber.complete();
-    this.unsubscriber = new Subject();
+    //    console.log("Unsubscribe");
+    this.unsubscriber.unsubscribe();
   }
 
   handleChange(change: Change): void {
     // No need to update anything
   }
-
-
 }
