@@ -1,13 +1,10 @@
 import { Component } from '@angular/core';
-import {BaseAppComponent} from "@dontcode/sandbox";
-import {dtcde} from "@dontcode/core";
-
+import { BaseAppComponent } from '@dontcode/sandbox';
+import { dtcde } from '@dontcode/core';
 
 @Component({
   selector: 'dontcode-tester-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  styleUrls: ['./app.component.scss'],
 })
-export class AppComponent extends BaseAppComponent{
-}
-
+export class AppComponent extends BaseAppComponent {}
