@@ -1,3 +1,3 @@
-require("fake-indexeddb/auto");
+require('fake-indexeddb/auto');
 
 import 'jest-preset-angular/setup-jest';
