@@ -1,3 +1,3 @@
 const nxPreset = require('@nrwl/jest/preset');
 
-module.exports = { ...nxPreset };
+module.exports = { ...nxPreset};
