@@ -1,39 +1,57 @@
-import {ChangeDetectorRef, Component, Inject, Injector, Input, OnInit, TemplateRef, ViewChild} from '@angular/core';
-import {Change, CommandProviderInterface, DontCodeModel, DontCodeModelPointer} from "@dontcode/core";
+import {
+  ChangeDetectorRef,
+  Component,
+  Inject,
+  Injector,
+  Input,
+  OnInit,
+  TemplateRef,
+  ViewChild,
+} from '@angular/core';
+import {
+  Change,
+  CommandProviderInterface,
+  DontCodeModel,
+  DontCodeModelPointer,
+} from '@dontcode/core';
 import {
   ComponentLoaderService,
   DynamicComponent,
   PluginBaseComponent,
   PossibleTemplateList,
-  TemplateList
-} from "@dontcode/plugin-common";
-import {FormBuilder, FormControl, FormGroup, Validators} from "@angular/forms";
+  TemplateList,
+} from '@dontcode/plugin-common';
+import {
+  FormBuilder,
+  FormControl,
+  FormGroup,
+  Validators,
+} from '@angular/forms';
 
 @Component({
   selector: 'dontcode-edit-entity',
   templateUrl: './edit-entity.component.html',
-  styleUrls: ['./edit-entity.component.scss']
+  styleUrls: ['./edit-entity.component.scss'],
 })
 export class EditEntityComponent extends PluginBaseComponent implements OnInit {
-
-  @Input("value") set _value(newval:any) {
+  @Input('value') set _value(newval: any) {
     this.value = newval;
 
-    if( this.value) {
-        // Transforms the stored values into field values
-      this.fields.forEach(field => {
+    if (this.value) {
+      // Transforms the stored values into field values
+      this.fields.forEach((field) => {
         if (field?.component?.managesFormControl()) {
           field.component.setValue(this.value[field.name]);
         } else {
           if (this.form) {
-            const singleVal:{[key:string]:any}={};
-            singleVal[field.name]=this.value[field.name];
-            this.form.patchValue(singleVal,{emitEvent:false});
-           }
+            const singleVal: { [key: string]: any } = {};
+            singleVal[field.name] = this.value[field.name];
+            this.form.patchValue(singleVal, { emitEvent: false });
+          }
         }
       });
     } else {
-      this.form?.reset({}, {emitEvent:false});
+      this.form?.reset({}, { emitEvent: false });
     }
   }
 
@@ -47,21 +65,24 @@ export class EditEntityComponent extends PluginBaseComponent implements OnInit {
 
   //formConfig = {};
 
-  constructor(private ref:ChangeDetectorRef, protected fb:FormBuilder, injector:Injector, componentLoader: ComponentLoaderService) {
+  constructor(
+    private ref: ChangeDetectorRef,
+    protected fb: FormBuilder,
+    injector: Injector,
+    componentLoader: ComponentLoaderService
+  ) {
     super(componentLoader, injector);
   }
 
   ngOnInit(): void {
-   this.form = this.fb.group({}, {updateOn:'blur'});
-    this.form.valueChanges.subscribe(change => {
+    this.form = this.fb.group({}, { updateOn: 'blur' });
+    this.form.valueChanges.subscribe((change) => {
       if (this.value) {
         for (const changeKey in change) {
-          if( change.hasOwnProperty(changeKey)) {
-            const field = this.fields.find(toSearch => {
-              if (toSearch.name===changeKey)
-                return true;
-              else
-                return false;
+          if (change.hasOwnProperty(changeKey)) {
+            const field = this.fields.find((toSearch) => {
+              if (toSearch.name === changeKey) return true;
+              else return false;
             });
             let newVal = change[changeKey];
             if (field?.component?.managesFormControl()) {
@@ -75,13 +96,22 @@ export class EditEntityComponent extends PluginBaseComponent implements OnInit {
     });
   }
 
-  initCommandFlow(provider: CommandProviderInterface, pointer: DontCodeModelPointer): any {
+  override initCommandFlow(
+    provider: CommandProviderInterface,
+    pointer: DontCodeModelPointer
+  ): any {
     //this.initing=true;
     super.initCommandFlow(provider, pointer);
 
-    if (!this.entityPointer)  throw new Error ('Cannot listen to changes without knowing a base position');
-    this.decomposeJsonToMultipleChanges (this.entityPointer, provider.getJsonAt(this.entityPointer.position)); // Dont provide a special handling for initial json, but emulate a list of changes
-    this.initChangeListening (true); // Listen to all changes occuring after entityPointer
+    if (!this.entityPointer)
+      throw new Error(
+        'Cannot listen to changes without knowing a base position'
+      );
+    this.decomposeJsonToMultipleChanges(
+      this.entityPointer,
+      provider.getJsonAt(this.entityPointer.position)
+    ); // Dont provide a special handling for initial json, but emulate a list of changes
+    this.initChangeListening(true); // Listen to all changes occuring after entityPointer
     //this.initing=false;
     this.rebuildForm();
   }
@@ -91,18 +121,23 @@ export class EditEntityComponent extends PluginBaseComponent implements OnInit {
    * @param change
    * @protected
    */
-  handleChange(change: Change) {
-    if (change.position!==this.entityPointer?.position) {
-
-      this.applyUpdatesToArrayAsync(this.fields, this.fieldsMap, change, null, (position, value) => {
-          return this.loadSubComponent(position, value).then(component => {
+  override handleChange(change: Change) {
+    if (change.position !== this.entityPointer?.position) {
+      this.applyUpdatesToArrayAsync(
+        this.fields,
+        this.fieldsMap,
+        change,
+        null,
+        (position, value) => {
+          return this.loadSubComponent(position, value).then((component) => {
             if (component) {
               component.setName(value.name);
               component.setForm(this.form);
             }
             return new FormElement(value.name, value.type, component);
           });
-        }, (elt, key, newVal) => {
+        },
+        (elt, key, newVal) => {
           switch (key) {
             case DontCodeModel.APP_FIELDS_NAME_NODE:
               elt.name = newVal;
@@ -112,12 +147,11 @@ export class EditEntityComponent extends PluginBaseComponent implements OnInit {
           }
           return true;
         }
-      ).then(updatedFields => {
+      ).then((updatedFields) => {
         this.fields = updatedFields;
         this.rebuildForm();
         this.ref.markForCheck();
         this.ref.detectChanges();
-
       });
     }
   }
@@ -127,36 +161,36 @@ export class EditEntityComponent extends PluginBaseComponent implements OnInit {
    * @private
    */
   private rebuildForm() {
-      // Updates the formgroup with new fields and remove old fields if necessary
+    // Updates the formgroup with new fields and remove old fields if necessary
     const toRemove = new Set<string>();
     // tslint:disable-next-line:forin
     for (const formKey in this.form.controls) {
       toRemove.add(formKey);
     }
 
-    this.fields.forEach(field => {
-      let val=null;
-      if((this.value)&&(this.value[field.name])) {
+    this.fields.forEach((field) => {
+      let val = null;
+      if (this.value && this.value[field.name]) {
         val = this.value[field.name];
       }
       toRemove.delete(field.name);
-      if( field.component)
-        field.component.setValue(val);
+      if (field.component) field.component.setValue(val);
 
       // Check if the component manages the FormControl itself or if it relies on us
-      if(!field.component?.managesFormControl())
-        this.form.registerControl(field.name, new FormControl(val, Validators.required));
-
+      if (!field.component?.managesFormControl())
+        this.form.registerControl(
+          field.name,
+          new FormControl(val, Validators.required)
+        );
     });
 
-    toRemove.forEach(key => {
+    toRemove.forEach((key) => {
       this.form.removeControl(key);
-    })
-
+    });
   }
 
   providesTemplates(): TemplateList {
-    return new TemplateList(null,null,null);
+    return new TemplateList(null, null, null);
   }
 
   canProvide(key?: string): PossibleTemplateList {
@@ -164,17 +198,18 @@ export class EditEntityComponent extends PluginBaseComponent implements OnInit {
   }
 
   templateOf(field: FormElement): TemplateRef<any> {
-    let ref= field.component?.providesTemplates(field.type).forFullEdit;
-    if( !ref)
-      ref = this.defaultTemplate;
+    let ref = field.component?.providesTemplates(field.type).forFullEdit;
+    if (!ref) ref = this.defaultTemplate;
 
     return ref;
   }
 
-  setForm(form: FormGroup) {
+  override setForm(form: FormGroup) {
     // Just ignore any form set
     if (form) {
-      throw new Error ("Trying to set a form to the Edit Entity component who already has one");
+      throw new Error(
+        'Trying to set a form to the Edit Entity component who already has one'
+      );
     }
   }
 }
@@ -182,12 +217,11 @@ export class EditEntityComponent extends PluginBaseComponent implements OnInit {
 class FormElement {
   type: string;
   name: string;
-  component: DynamicComponent|null;
+  component: DynamicComponent | null;
 
-  constructor(name:string,type:string, component:DynamicComponent|null) {
-    this.name=name;
-    this.type=type;
+  constructor(name: string, type: string, component: DynamicComponent | null) {
+    this.name = name;
+    this.type = type;
     this.component = component;
   }
-
 }
