@@ -9,10 +9,13 @@ module.exports = {
     },
   },
   coverageDirectory: '../../coverage/libs/common',
+  transform: {
+    '^.+\\.(ts|mjs|js|html)$': 'jest-preset-angular',
+  },
+  transformIgnorePatterns: ['node_modules/(?!.*\\.mjs$)'],
   snapshotSerializers: [
     'jest-preset-angular/build/serializers/no-ng-attributes',
     'jest-preset-angular/build/serializers/ng-snapshot',
     'jest-preset-angular/build/serializers/html-comment',
-  ],
-  transform: { '^.+\\.(ts|js|html)$': 'jest-preset-angular' },
+  ]
 };
