@@ -1,22 +1,30 @@
-import {Component, Injector, OnDestroy} from "@angular/core";
-import {Change, CommandProviderInterface, DontCodeModelPointer, PreviewHandler} from "@dontcode/core";
-import {ComponentLoaderService} from "../common-dynamic/component-loader.service";
-import {AbstractDynamicLoaderComponent} from "./abstract-dynamic-loader-component";
-import {PluginHandlerHelper} from "../common-handler/plugin-handler-helper";
-import {Subscription} from "rxjs";
+import { Component, Injector, OnDestroy } from '@angular/core';
+import {
+  Change,
+  CommandProviderInterface,
+  DontCodeModelPointer,
+  PreviewHandler,
+} from '@dontcode/core';
+import { ComponentLoaderService } from '../common-dynamic/component-loader.service';
+import { AbstractDynamicLoaderComponent } from './abstract-dynamic-loader-component';
+import { PluginHandlerHelper } from '../common-handler/plugin-handler-helper';
+import { Subscription } from 'rxjs';
 
 /**
  * A component that can be loaded by the framework, load subcomponents, listen to model changes, and so on...
  * Usually provided by plugins and run by the framework
  */
-@Component({template: ''})
-export abstract class PluginBaseComponent extends AbstractDynamicLoaderComponent implements PreviewHandler, OnDestroy {
+@Component({ template: '' })
+export abstract class PluginBaseComponent
+  extends AbstractDynamicLoaderComponent
+  implements PreviewHandler, OnDestroy
+{
   protected subscriptions = new Subscription();
   protected pluginHelper = new PluginHandlerHelper();
-  entityPointer: DontCodeModelPointer | null = null
+  entityPointer: DontCodeModelPointer | null = null;
   protected provider: CommandProviderInterface | null = null;
 
-  constructor(loader: ComponentLoaderService, injector:Injector) {
+  constructor(loader: ComponentLoaderService, injector: Injector) {
     super(loader, injector);
   }
 
@@ -29,13 +37,16 @@ export abstract class PluginBaseComponent extends AbstractDynamicLoaderComponent
     this.subscriptions.unsubscribe();
   }
 
-  initCommandFlow(provider: CommandProviderInterface, pointer: DontCodeModelPointer): any {
+  initCommandFlow(
+    provider: CommandProviderInterface,
+    pointer: DontCodeModelPointer
+  ): any {
     this.entityPointer = pointer;
     this.provider = provider;
     this.pluginHelper.initCommandFlow(provider, pointer, this);
   }
 
-  protected initChangeListening(subElement?:boolean) {
+  protected initChangeListening(subElement?: boolean) {
     this.pluginHelper.initChangeListening(subElement);
   }
 
@@ -45,7 +56,10 @@ export abstract class PluginBaseComponent extends AbstractDynamicLoaderComponent
    * This will avoid to duplicate code (first time when a complete json is sent, second when subelements are sent)
    * @protected
    */
-  protected decomposeJsonToMultipleChanges(pointer: DontCodeModelPointer, json: any): void {
+  protected decomposeJsonToMultipleChanges(
+    pointer: DontCodeModelPointer,
+    json: any
+  ): void {
     this.pluginHelper.decomposeJsonToMultipleChanges(pointer, json);
   }
 
@@ -54,11 +68,10 @@ export abstract class PluginBaseComponent extends AbstractDynamicLoaderComponent
    * @param change
    * @param key
    */
-  decodeStringField(change: Change, key: string): string|undefined {
+  decodeStringField(change: Change, key: string): string | undefined {
     if (change.pointer?.lastElement === key) {
       return change.value;
-    } else
-      return undefined;
+    } else return undefined;
   }
 
   /**
@@ -70,10 +83,23 @@ export abstract class PluginBaseComponent extends AbstractDynamicLoaderComponent
    * @param transform
    * @private
    */
-  applyUpdatesToArray<T>(target: T[], targetMap: Map<string, number>, change: Change, property: string, transform: (position: DontCodeModelPointer, item: any) => T, applyProperty?: (target: T, key: string, value: any) => boolean): Promise<T[]> {
-    return this.applyUpdatesToArrayAsync(target, targetMap, change, property, (key, item) => {
-      return Promise.resolve( transform(key, item));
-    } );
+  applyUpdatesToArray<T>(
+    target: T[],
+    targetMap: Map<string, number>,
+    change: Change,
+    property: string,
+    transform: (position: DontCodeModelPointer, item: any) => T,
+    applyProperty?: (target: T, key: string, value: any) => boolean
+  ): Promise<T[]> {
+    return this.applyUpdatesToArrayAsync(
+      target,
+      targetMap,
+      change,
+      property,
+      (key, item) => {
+        return Promise.resolve(transform(key, item));
+      }
+    );
   }
 
   /**
@@ -85,17 +111,28 @@ export abstract class PluginBaseComponent extends AbstractDynamicLoaderComponent
    * @param transform
    * @private
    */
-  applyUpdatesToArrayAsync<T>(target: T[], targetMap: Map<string, number>, change: Change, property: string|null, transform: (position: DontCodeModelPointer, item: any) => Promise<T>, applyProperty?: (target: T, key: string|null, value: any) => boolean): Promise<T[]> {
-    return this.pluginHelper.applyUpdatesToArrayAsync(target, targetMap, change, property, transform, applyProperty);
+  applyUpdatesToArrayAsync<T>(
+    target: T[],
+    targetMap: Map<string, number>,
+    change: Change,
+    property: string | null,
+    transform: (position: DontCodeModelPointer, item: any) => Promise<T>,
+    applyProperty?: (target: T, key: string | null, value: any) => boolean
+  ): Promise<T[]> {
+    return this.pluginHelper.applyUpdatesToArrayAsync(
+      target,
+      targetMap,
+      change,
+      property,
+      transform,
+      applyProperty
+    );
   }
 
-    /**
+  /**
    * This is where components react to changes received
    * @param change
    * @protected
    */
-  handleChange(change: Change) {
-
-  }
-
+  handleChange(change: Change) {}
 }
