@@ -1,22 +1,40 @@
-import {Change, CommandProviderInterface, DontCodeModel, DontCodeModelPointer, DontCodeStoreProvider, dtcde} from "@dontcode/core";
-import {ChangeDetectionStrategy, ChangeDetectorRef, Component, Injector, OnInit, TemplateRef} from "@angular/core";
+import {
+  Change,
+  CommandProviderInterface,
+  DontCodeModel,
+  DontCodeModelPointer,
+  DontCodeStoreProvider,
+  dtcde,
+} from '@dontcode/core';
+import {
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  Injector,
+  OnInit,
+  TemplateRef,
+} from '@angular/core';
 import {
   ComponentLoaderService,
   DynamicComponent,
   PluginBaseComponent,
   PossibleTemplateList,
-  TemplateList
-} from "@dontcode/plugin-common";
-import {FormBuilder, FormControl, FormGroup, Validators} from "@angular/forms";
+  TemplateList,
+} from '@dontcode/plugin-common';
+import {
+  FormBuilder,
+  FormControl,
+  FormGroup,
+  Validators,
+} from '@angular/forms';
 
 @Component({
   selector: 'dontcode-sandbox-default-viewer',
   templateUrl: './default-viewer.component.html',
   styleUrls: ['./default-viewer.component.css'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class DefaultViewerComponent extends PluginBaseComponent {
-
   fields = new Array<Field>();
   fieldsMap = new Map<string, number>();
 
@@ -24,23 +42,33 @@ export class DefaultViewerComponent extends PluginBaseComponent {
 
   store?: DontCodeStoreProvider;
 
-  constructor(loader: ComponentLoaderService, injector: Injector, protected ref: ChangeDetectorRef, protected fb: FormBuilder) {
+  constructor(
+    loader: ComponentLoaderService,
+    injector: Injector,
+    protected ref: ChangeDetectorRef,
+    protected fb: FormBuilder
+  ) {
     super(loader, injector);
-    this.setForm( this.fb.group({}, {updateOn: 'blur'}));
+    this.setForm(this.fb.group({}, { updateOn: 'blur' }));
   }
 
-  initCommandFlow(provider: CommandProviderInterface, pointer: DontCodeModelPointer) {
+  override initCommandFlow(
+    provider: CommandProviderInterface,
+    pointer: DontCodeModelPointer
+  ) {
     super.initCommandFlow(provider, pointer);
     if (this.isEntity()) {
-      this.decomposeJsonToMultipleChanges(pointer, provider.getJsonAt(pointer.position));
+      this.decomposeJsonToMultipleChanges(
+        pointer,
+        provider.getJsonAt(pointer.position)
+      );
       this.store = dtcde.getStoreManager().getProvider(pointer.position);
-      if (this.store!=null) {
-        this.store.loadEntity(pointer.position,null).then ( val => {
-          this.setValue( val);
+      if (this.store != null) {
+        this.store.loadEntity(pointer.position, null).then((val) => {
+          this.setValue(val);
           this.rebuildForm();
           this.ref.detectChanges();
-          }
-        );
+        });
       }
       this.initChangeListening();
       this.rebuildForm();
@@ -48,25 +76,31 @@ export class DefaultViewerComponent extends PluginBaseComponent {
     this.ref.detectChanges();
   }
 
-
-  handleChange(change: Change) {
+  override handleChange(change: Change) {
     super.handleChange(change);
 
     if (this.entityPointer) {
       if (change?.pointer?.positionInSchema === DontCodeModel.APP_FIELDS) {
-        this.applyUpdatesToArrayAsync(this.fields, this.fieldsMap, change, 'fields', (position, value) => {
-          return this.loadSubField(value.type, value.name, null).then(component => {
-
-            const ret = new Field(value.name, value.type);
-            if (component) {
-              // Keep the component only if it provides the view template
-              if (component.canProvide(value.type).forInlineView) {
-                ret.component = component;
+        this.applyUpdatesToArrayAsync(
+          this.fields,
+          this.fieldsMap,
+          change,
+          'fields',
+          (position, value) => {
+            return this.loadSubField(value.type, value.name, null).then(
+              (component) => {
+                const ret = new Field(value.name, value.type);
+                if (component) {
+                  // Keep the component only if it provides the view template
+                  if (component.canProvide(value.type).forInlineView) {
+                    ret.component = component;
+                  }
+                }
+                return ret;
               }
-            }
-            return ret;
-          });
-        }).then(updatedColumns => {
+            );
+          }
+        ).then((updatedColumns) => {
           this.fields = updatedColumns;
           this.rebuildForm();
           //  this.reloadData ();
@@ -80,15 +114,13 @@ export class DefaultViewerComponent extends PluginBaseComponent {
         this.ref.detectChanges();
       }
     }
-
   }
 
   templateOf(col: Field, value: any): TemplateRef<any> {
     if (col.component) {
       col.component.setValue(value);
       const ref = col.component.providesTemplates(col.type).forInlineView;
-      if (ref)
-        return ref;
+      if (ref) return ref;
     }
     throw new Error('No component or template to display ' + col.type);
   }
@@ -97,8 +129,7 @@ export class DefaultViewerComponent extends PluginBaseComponent {
     if (col.component) {
       col.component.setValue(value);
       const ref = col.component.providesTemplates(col.type).forFullEdit;
-      if (ref)
-        return ref;
+      if (ref) return ref;
     }
     throw new Error('No component or template to display ' + col.type);
   }
@@ -130,44 +161,43 @@ export class DefaultViewerComponent extends PluginBaseComponent {
         toRemove.add(formKey);
       }
 
-      this.fields.forEach(field => {
+      this.fields.forEach((field) => {
         let val = null;
-        if ((this.value) && (this.value[field.name])) {
+        if (this.value && this.value[field.name]) {
           val = this.value[field.name];
         }
         toRemove.delete(field.name);
-        if (field.component)
-          field.component.setValue(val);
+        if (field.component) field.component.setValue(val);
 
         // Check if the component manages the FormControl itself or if it relies on us
         if (!field.component?.managesFormControl())
-          this.form.setControl(field.name, new FormControl(val, Validators.required));
-
+          this.form.setControl(
+            field.name,
+            new FormControl(val, Validators.required)
+          );
       });
 
-      toRemove.forEach(key => {
+      toRemove.forEach((key) => {
         this.form.removeControl(key);
       });
     }
   }
 
-  getData(fieldName: string): any|undefined {
+  getData(fieldName: string): any | undefined {
     if (this.form.controls[fieldName])
       return this.form.controls[fieldName].value;
     else return undefined;
   }
-
 }
 
 class Field {
-  name:string;
-  type:string;
-  component: DynamicComponent|null;
+  name: string;
+  type: string;
+  component: DynamicComponent | null;
 
   constructor(name: string, type: string) {
     this.name = name;
     this.type = type;
-    this.component=null;
+    this.component = null;
   }
-
 }
